+++ backend/canisters/community/impl/src/model/local_user_index_event_batch.rs
@@ -1,13 +1,13 @@
 use crate::{can_borrow_state, run_regular_jobs};
 use local_user_index_canister::CommunityEvent;
 use timer_job_queues::{TimerJobItem, timer_job_batch};
-use types::{CanisterId, IdempotentEnvelope};
-use utils::canister::should_retry_failed_c2c_call;
+use types::{CanisterId, IdempotentEnvelope, Milliseconds};
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 timer_job_batch!(LocalUserIndexEventBatch, CanisterId, IdempotentEnvelope<CommunityEvent>, 10);
 
 impl TimerJobItem for LocalUserIndexEventBatch {
-    async fn process(&self) -> Result<(), bool> {
+    async fn process(&self) -> Result<(), Option<Milliseconds>> {
         if can_borrow_state() {
             run_regular_jobs();
         }
@@ -23,8 +23,8 @@ impl TimerJobItem for LocalUserIndexEventBatch {
         match response {
             Ok(local_user_index_canister::c2c_community_canister::Response::Success) => Ok(()),
             Err(error) => {
-                let retry = should_retry_failed_c2c_call(error.reject_code(), error.message());
-                Err(retry)
+                let delay_if_should_retry = delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message());
+                Err(delay_if_should_retry)
             }
         }
     }