+++ backend/canisters/group_index/impl/src/model/local_index_event_batch.rs
@@ -1,12 +1,12 @@
 use crate::LocalIndexEvent;
 use timer_job_queues::{TimerJobItem, grouped_timer_job_batch};
-use types::{CanisterId, IdempotentEnvelope};
-use utils::canister::should_retry_failed_c2c_call;
+use types::{CanisterId, IdempotentEnvelope, Milliseconds};
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 grouped_timer_job_batch!(LocalIndexEventBatch, CanisterId, IdempotentEnvelope<LocalIndexEvent>, 1000);
 
 impl TimerJobItem for LocalIndexEventBatch {
-    async fn process(&self) -> Result<(), bool> {
+    async fn process(&self) -> Result<(), Option<Milliseconds>> {
         let response = local_user_index_canister_c2c_client::c2c_group_index(
             self.key,
             &local_user_index_canister::c2c_group_index::Args {
@@ -18,8 +18,8 @@ impl TimerJobItem for LocalIndexEventBatch {
         match response {
             Ok(_) => Ok(()),
             Err(error) => {
-                let retry = should_retry_failed_c2c_call(error.reject_code(), error.message());
-                Err(retry)
+                let delay_if_should_retry = delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message());
+                Err(delay_if_should_retry)
             }
         }
     }