+++ backend/canisters/proposals_bot/impl/src/jobs/update_finished_proposals.rs
@@ -7,7 +7,7 @@ use std::cell::Cell;
 use std::time::Duration;
 use tracing::trace;
 use types::{C2CError, CanisterId, Proposal, ProposalId};
-use utils::canister::should_retry_failed_c2c_call;
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 thread_local! {
     static TIMER_ID: Cell<Option<TimerId>> = Cell::default();
@@ -67,7 +67,7 @@ async fn process_proposal(governance_canister_id: CanisterId, proposal_id: Propo
                     .finished_proposals_to_process
                     .push_back((governance_canister_id, proposal_id));
 
-                if !should_retry_failed_c2c_call(error.reject_code(), error.message()) {
+                if delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message()).is_none() {
                     state.data.nervous_systems.mark_disabled(&governance_canister_id);
                 }