+++ backend/canisters/user_index/impl/src/jobs/sync_events_to_local_user_index_canisters.rs
@@ -4,7 +4,7 @@ use std::cell::Cell;
 use std::time::Duration;
 use tracing::trace;
 use types::CanisterId;
-use utils::canister::should_retry_failed_c2c_call;
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 thread_local! {
     static TIMER_ID: Cell<Option<TimerId>> = Cell::default();
@@ -62,7 +62,7 @@ async fn process_batch(batch: Vec<(CanisterId, Vec<LocalUserIndexEvent>)>) {
 async fn sync_events(canister_id: CanisterId, events: Vec<LocalUserIndexEvent>) {
     let args = local_user_index_canister::c2c_notify_user_index_events::Args { events: events.clone() };
     if let Err(error) = local_user_index_canister_c2c_client::c2c_notify_user_index_events(canister_id, &args).await {
-        if should_retry_failed_c2c_call(error.reject_code(), error.message()) {
+        if delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message()).is_some() {
             mutate_state(|state| {
                 state
                     .data