+++ backend/libraries/fire_and_forget_handler/src/lib.rs
@@ -10,7 +10,7 @@ use std::sync::Mutex;
 use std::time::Duration;
 use tracing::trace;
 use types::{CanisterId, TimestampMillis};
-use utils::canister::should_retry_failed_c2c_call;
+use utils::canister::delay_if_should_retry_failed_c2c_call;
 
 pub struct FireAndForgetHandler {
     inner: Rc<Mutex<FireAndForgetHandlerInner>>,
@@ -55,7 +55,10 @@ impl FireAndForgetHandler {
                 calls.in_progress.retain(|id| *id != call.id);
 
                 match result {
-                    Err(error) if should_retry_failed_c2c_call(error.reject_code(), error.message()) && call.attempt < 50 => {
+                    Err(error)
+                        if delay_if_should_retry_failed_c2c_call(error.reject_code(), error.message()).is_some()
+                            && call.attempt < 50 =>
+                    {
                         call.attempt += 1;
                         let now = canister_time::now_millis();
                         let due = now + (u64::from(call.attempt) * SECOND_IN_MS);